chore: Update dependencies and started to refactor code structure
