+++ Client/package.json
@@ -16,6 +16,7 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
     "lucide-vue-next": "^0.320.0",
+    "pinia": "^2.1.7",
     "radix-vue": "^1.3.2",
     "tailwind-merge": "^2.2.1",
     "tailwindcss-animate": "^1.0.7",
