+++ Client/src/components/header/Logo.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts"></script>
+
+<template>
+  <div class="flex items-center">
+    <img class="size-20 mr-2" src="/squi.svg" alt="logo" />
+  </div>
+</template>
